debut algo

+++ libft/ft_atoi.c
@@ -6,17 +6,17 @@
 /*   By: msainton <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/26 14:43:49 by msainton          #+#    #+#             */
-/*   Updated: 2021/10/07 19:04:14 by msainton         ###   ########.fr       */
+/*   Updated: 2021/11/04 16:58:02 by msainton         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_atoi(const char *str)
+long int	ft_atoi(const char *str)
 {
-	int	a;
-	int	sign;
-	int	nbr;
+	int			a;
+	int			sign;
+	long int	nbr;
 
 	a = 0;
 	sign = 1;
